reducer/requirements.txt
# Function dependencies
google-cloud-storage
